+++ src/main/kotlin/io/github/kscripting/kscript/util/ShellUtils.kt
@@ -15,7 +15,7 @@ object ShellUtils {
     }
 
     fun guessPosixKotlinHome(osType: OsType): String? {
-        val kotlinHome = ShellExecutor.eval(osType, "KOTLIN_RUNNER=1 JAVACMD=echo kotlinc").stdout.run {
+        val kotlinHome = ShellExecutor.evalAndGobble(osType, "KOTLIN_RUNNER=1 JAVACMD=echo kotlinc").stdout.run {
             "kotlin.home=([^\\s]*)".toRegex().find(this)?.groups?.get(1)?.value
         } ?: return null
 
@@ -59,7 +59,7 @@ object ShellUtils {
         return null
     }
 
-    fun which(osType: OsType, command: String, envAdjuster: EnvAdjuster = {}): List<String> = ShellExecutor.eval(
+    fun which(osType: OsType, command: String, envAdjuster: EnvAdjuster = {}): List<String> = ShellExecutor.evalAndGobble(
         osType, "${if (osType == OsType.WINDOWS) "where" else "which"} $command", null, envAdjuster
     ).stdout.trim().lines()